+++ src/index.ts
@@ -1,3 +1,4 @@
+import { Options } from './api'
 import { Client } from './client'
 import { QueryBuilder } from './query_builder'
 
@@ -11,19 +12,27 @@ const throwParamRequiredError = (param: string) => {
  * @param {string} endpoint
  * @param {string} apiKey
  * @param {Object} params
+ * @param {Object} apiOptions
  * @returns {function(): QueryBuilder}
  */
 export default function SearchApplicationClient(
   applicationName: string,
   endpoint: string,
   apiKey: string,
-  params?: Record<string, any>
+  params?: Record<string, any>,
+  apiOptions?: Options
 ): () => QueryBuilder {
   if (!applicationName) throwParamRequiredError('applicationName')
   if (!endpoint) throwParamRequiredError('endpoint')
   if (!apiKey) throwParamRequiredError('apiKey')
 
-  const client = new Client(applicationName, endpoint, apiKey, params)
+  const client = new Client({
+    applicationName,
+    endpoint,
+    apiKey,
+    baseParams: params,
+    apiOptions,
+  })
 
   return () => client.initQuery()
 }