+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # @elastic/search-application-client
 
+## 8.9.0-2
+
+### Patch Changes
+
+- f4b43f3: Update caching options
+- 36dd7dc: Fix empty facet issue
+
 ## 1.0.1
 
 ### Patch Changes